fix: misleading error messages (#98)
* fix: misleading error messages

Signed-off-by: Nikolay Manchev <nikolay.manchev@limechain.tech>

* fix: incorrect error message was shown

Due to the relation between Schedule Time(minutely, hourly, dayly, etc) and the Every field
we had to do some shenanigans to estimate whether the current field is Every or not. This
shenanigans were failing for I want to allocate field and was showing the wrong error message

Signed-off-by: Nikolay Manchev <nikolay.manchev@limechain.tech>

* fix: proper fix for error messages tangling

The current approach to validating Every field based on Time field was brittle and was bringing a lot of headaches with it. Implemented a more robust solution using if/else

Signed-off-by: Nikolay Manchev <nikolay.manchev@limechain.tech>

---------

Signed-off-by: Nikolay Manchev <nikolay.manchev@limechain.tech>